UI: Hide labels from MatCaps
They were added as placeholder to show something until they lazy-load.
But since the load is fast and it's hard to fit the text, and their name
is displayed already in the tooltip, we can do without for now.

In the future when we have insta-tooltips we should make them use this.